# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blockchain Freelance Marketplace API",
    "version": "1.0.0",
    "description": "A decentralized freelance marketplace with AI skill matching and blockchain payments",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MilestoneCompletionResult": {
        "type": "object",
        "properties": {
          "milestoneId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "submitted"
            ]
          },
          "notificationSent": {
            "type": "boolean"
          }
        }
      },
      "MilestoneApprovalResult": {
        "type": "object",
        "properties": {
          "milestoneId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "approved"
            ]
          },
          "paymentReleased": {
            "type": "boolean"
          },
          "transactionHash": {
            "type": "string"
          },
          "contractCompleted": {
            "type": "boolean"
          }
        }
      },
      "MilestoneDisputeResult": {
        "type": "object",
        "properties": {
          "milestoneId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "disputed"
            ]
          },
          "disputeId": {
            "type": "string"
          },
          "disputeCreated": {
            "type": "boolean"
          }
        }
      },
      "ContractPaymentStatus": {
        "type": "object",
        "properties": {
          "contractId": {
            "type": "string"
          },
          "escrowAddress": {
            "type": "string"
          },
          "totalAmount": {
            "type": "number"
          },
          "releasedAmount": {
            "type": "number"
          },
          "pendingAmount": {
            "type": "number"
          },
          "milestones": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "status": {
                  "type": "string"
                }
              }
            }
          },
          "contractStatus": {
            "type": "string"
          }
        }
      },
      "DisputeRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for disputing the milestone"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/payments/milestones/{milestoneId}/complete": {
      "post": {
        "summary": "Mark milestone as complete",
        "description": "Freelancer marks a milestone as complete, triggering employer notification",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "milestoneId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The milestone ID (UUID)"
          },
          {
            "in": "query",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone marked as complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneCompletionResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or UUID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract or milestone not found"
          }
        }
      }
    },
    "/api/payments/milestones/{milestoneId}/approve": {
      "post": {
        "summary": "Approve milestone completion",
        "description": "Employer approves milestone completion, triggering payment release",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "milestoneId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The milestone ID (UUID)"
          },
          {
            "in": "query",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone approved and payment released",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneApprovalResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or UUID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract or milestone not found"
          }
        }
      }
    },
    "/api/payments/milestones/{milestoneId}/dispute": {
      "post": {
        "summary": "Dispute milestone",
        "description": "Either party disputes a milestone, locking funds and creating a dispute record",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "milestoneId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The milestone ID (UUID)"
          },
          {
            "in": "query",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The contract ID (UUID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispute created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneDisputeResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or UUID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract or milestone not found"
          }
        }
      }
    },
    "/api/payments/contracts/{contractId}/status": {
      "get": {
        "summary": "Get contract payment status",
        "description": "Get detailed payment status for a contract including milestone statuses",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The contract ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Contract payment status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractPaymentStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract not found"
          }
        }
      }
    },
...
  },
...
}